+++ Runner/src/NPC.h
@@ -45,7 +45,7 @@ public:
 	SMG weapon;
 
 	virtual sf::Vector2f getPoint(int i) const;
-	virtual size_t getPointCount() const;
+	virtual unsigned int getPointCount() const;
 	virtual sf::Transform getTransform() const;
 	virtual float getMinDistance() const;
 };
