+++ Runner/src/Player.h
@@ -54,7 +54,7 @@ public:
 
 private:
 	virtual sf::Vector2f getPoint(int i) const;
-	virtual size_t getPointCount() const;
+	virtual unsigned int getPointCount() const;
 	virtual sf::Transform getTransform() const;
 	virtual float getMinDistance() const;
 };
